requirements.txt
torch
transformers
sentence-transformers
numpy
pandas
openai
python-dotenv
pyyaml
faiss-cpu
scipy
matplotlib
seaborn